BAU: Read config from environment variables
In my attempts to try and 12-factorise the middleware as much as
possible, I want to get the config generator to read all the stuff it
needs from environment variables. It now does this if you pass the
`--env` flag to `generate`. The environment variables are the same as
the config file keys (which have changed a bit) but with `MW_` prefixed.

The code has been refactored to make it slightly more readable. All the
Docker stuff has been revamped slightly to make `pre-commit.sh` able to
test generating the middleware config against both a YAML config file and env vars.