EID-1159 Activate access logs for middleware
We don't currently collect the access logs for the middleware. They
would be useful.

This updates the template for the application.properties to enable this.

The log file will be rotated when the date changes.

We also defined the remote ip and protocol headers - this allows the
access logs to use the IP address specified in the X-Forwarded-For
header of the request coming from the ALB. Without it, we'd only log the
ALB's IP address. Hopefully this should be enough to get the client IP.
If not, we should be able to set some trusted IPs to ignore.